Refactored out plotting out to STLPlotter and changed test dependencies to reduce the compile dependency on JFreeChart
