+++ pom.xml
@@ -34,6 +34,7 @@
             <groupId>org.jfree</groupId>
             <artifactId>jfreechart</artifactId>
             <version>1.0.19</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
